feat: Implement UserPage with user information editing and logout functionality refactor: Update App component routes and remove unused components fix: Update Breadcrumbs to include new routes for Privacy Policy and Refund Policy chore: Remove useDocumentTitle hook and adjust product query handling feat: Add loading state to ProductPage and integrate Breadcrumbs component
