+++ fec/src/components/RadioSelector.tsx
@@ -8,15 +8,14 @@ type RadioSelectorProps = {
   selectedIndex?: number;
   required?: boolean;
   disabled?: boolean;
-  error?: string;
 };
 
 const RadioSelector = (props: RadioSelectorProps) => (
-  <div className={`inline-flex flex-wrap gap-4 ${props.className ?? ""}`}>
+  <div className={`group inline-flex flex-wrap gap-4 ${props.className ?? ""}`}>
     {props.options.map((option, index) => (
       <div key={option} className="flex items-center gap-1">
         <input
-          className="peer accent-red-500 nth-[n+2]:ms-4 focus:outline-red-500"
+          className="group accent-red-500 focus:outline-red-500 nth-[n+2]:ms-4"
           type="radio"
           id={props.groupName + index}
           name={props.groupName}
@@ -28,8 +27,6 @@ const RadioSelector = (props: RadioSelectorProps) => (
         <label htmlFor={props.groupName + index}>{option}</label>
       </div>
     ))}
-
-    <span className="ms-auto hidden text-red-500 peer-invalid:block">{props.error}</span>
   </div>
 );
 export default memo(RadioSelector);
