+++ fec/src/pages/PageNotFound.tsx
@@ -1,0 +1,12 @@
+import { memo } from "react";
+
+const PageNotFound = () => {
+  return (
+    <div className="flex h-screen flex-col items-center justify-center bg-gray-100">
+      <h1 className="text-6xl font-bold text-gray-800">404</h1>
+      <p className="mt-4 text-lg text-gray-600">Page Not Found</p>
+    </div>
+  );
+};
+
+export default memo(PageNotFound);
